+++ apps/dashboard/components/landing-page/Header.jsx
@@ -224,9 +224,9 @@ export function Header() {
             </div>
             <div className="flex gap-x-5 items-center md:gap-x-8">
               {/* <div className="hidden md:block">
-              <NavLink href="https://http://dashboard.chatsappai.com/">Sign in</NavLink>
+              <NavLink href="https://dashboard.chatsappai.com/">Sign in</NavLink>
             </div> */}
-              {/* <Button target="_blank" href="https://http://dashboard.chatsappai.com//signin" color="blue">
+              {/* <Button target="_blank" href="https://dashboard.chatsappai.com//signin" color="blue">
               <span>
                 Sign In
               </span>