+++ apps/dashboard/widgets/zapier/triggers/list_agents.js
@@ -4,7 +4,7 @@ zapier.tools.env.inject();
 
 const baseApiUrl = process.env.API_URL
   ? process.env.API_URL
-  : 'https://http://dashboard.chatsappai.com/';
+  : 'https://dashboard.chatsappai.com/';
 
 module.exports = {
   operation: {