+++ apps/lp/next.config.js
@@ -50,7 +50,7 @@ const nextConfig = {
         has: [
           {
             type: 'host',
-            value: 'http://dashboard.chatsappai.com/',
+            value: 'https://dashboard.chatsappai.com/',
           },
         ],
       },