+++ packages/embeds/esbuild.config.mjs
@@ -29,7 +29,7 @@ let ctx = await esbuild.context({
           'process.env.NEXT_PUBLIC_ASSETS_BASE_URL':
             '"https://cdn.jsdelivr.net/npm/@chaindesk/embeds@latest"',
           'process.env.NEXT_PUBLIC_DASHBOARD_URL':
-            '"https://http://dashboard.chatsappai.com/"',
+            '"https://dashboard.chatsappai.com/"',
         }
       : {}),
   },