+++ packages/lib/config.ts
@@ -160,7 +160,7 @@ export const ModelConfig: Record<
   },
 };
 
-export const appUrl = 'https://http://dashboard.chatsappai.com/';
+export const appUrl = 'https://dashboard.chatsappai.com/';
 export const apiUrl = 'https://api.chatsappai.com';
 // export const appUrl = process.env.NEXT_PUBLIC_DASHBOARD_URL as string;
 // export const apiUrl = process.env.NEXT_PUBLIC_API_URL as string;